test(desktop): check mounted names against the packaged closure
`require.resolve` answered a weaker question than the test's comment
claimed. Under vitest, NODE_PATH ends in `node_modules/.pnpm/node_modules`
- pnpm's hidden hoist, which holds the entire store - so the check was
"is this package anywhere on disk", not "will the packaged app have it".
Measured: mounting `@deepseek-ai/dsh-authorization`, the one installed
harness package outside the production closure, passed.

The closure walk the packaged-dependency test already performs is the
right answer, so it moves next to `installedHarnessPackages()` and both
tests read it. Same probe now fails, naming the package.

Also drops the `agent-instructions` override. Every key in it was inert:
dsh-web-app disables the global entry because each agent preset mounts
its own, and the preset's row states no `dshHome`, so the user-global
AGENTS.md resolves from the real DSH home either way. It read as live
configuration and was not.

What that override was reaching for is true but not configurable: the
home is `~/.pawwork/dsh`, so a file left in `~/.pawwork/` is silently
unread. Hoisting the home is what fixes it, and the home is one level
down only because v1 still owns `~/.pawwork` on maint/v1 - so the note
now sits beside the constant that will move when v1 retires.
